==> Step0_Data_Extraction.ipynb <==
# Script purpose

Created on 01/03/2023 by Claire

The output created by this script is used in the Step1_Regex_Data_Cleaning script. 

This script will do the following things:
- Issue queries from the sqlite file to filter either on labelled or unlabelled data
- Do some light cleaning, e.g. to split up the attachment types
- Save the extracted data files as pickle files for further analysis 

%load_ext autoreload
%autoreload 2

import pandas as pd
import pickle

# try and import my own functions:
import sys
import ipynb
import re
sys.path.append('/project/Xelix_Project/utils')

from ipynb.fs.full.Data_Extraction_Functions import return_labelled_data, return_unlabelled_data

# f_data = "/project/Xelix_Project/data/email_extractor.sqlite"
f_data = "data/email_extractor.sqlite"

# Extract labelled emails

These are the emails where the category slug needs to be 'not null' 

%%time
labelled_data = return_labelled_data(f_data)

print(len(labelled_data))
labelled_data.head()

labelled_data['category_slug'].value_counts()

print(sum(labelled_data.duplicated()))

# Extract unlabelled emails

Set a limit to extract, because large amount of emails

%%time
# limit = 15000    # how many emails to return
# start_from = 20000  # which rows to start from, e.g. 0 is from the start, 1000 is to skip the first 1000 rows

# limit = 20000    # how many emails to return
# start_from = 20000  # which rows to start from, e.g. 0 is from the start, 1000 is to skip the first 1000 rows

limit = 50000    # how many emails to return
start_from = 250000  # which rows to start from, e.g. 0 is from the start, 1000 is to skip the first 1000 rows

unlabelled_data = return_unlabelled_data(f_data, start_from, limit)

print(len(unlabelled_data))
unlabelled_data.head()

unlabelled_data['category_slug'].value_counts()

print(sum(unlabelled_data.duplicated()))

# Initial cleaning of data

This will:
- remove duplicate entries from the attachments column
- remove images and hyperlinks from the attachment types
- add new feature whether there is an attachment with the email or not after cleaning

from ipynb.fs.full.Regex_html_Functions import remove_duplicates, clean_attachment_types

# for the labelled data
labelled_data['attachments'] = labelled_data['attachments'].apply(remove_duplicates)
labelled_data['attachments'] = labelled_data['attachments'].apply(clean_attachment_types)
labelled_data['has_attachment'] = [True if val is not None else False for val in labelled_data['attachments']]

# for the unlabelled data
unlabelled_data['attachments'] = unlabelled_data['attachments'].apply(remove_duplicates)
unlabelled_data['attachments'] = unlabelled_data['attachments'].apply(clean_attachment_types)
unlabelled_data['has_attachment'] = [True if val is not None else False for val in unlabelled_data['attachments']]

# HTML parsing

This will turn the email text data into parsed plain text data from html if needed

from ipynb.fs.full.Regex_html_Functions import text_from_html

%%time
labelled_data["html_body_as_text"] = labelled_data["html_body"].map(text_from_html)
labelled_data["body"] = labelled_data["plain_text_body"]
labelled_data.loc[labelled_data["body"].isnull(), "body"] = labelled_data.loc[labelled_data["body"].isnull(), "html_body_as_text"]

%%time
unlabelled_data["html_body_as_text"] = unlabelled_data["html_body"].map(text_from_html)
unlabelled_data["body"] = unlabelled_data["plain_text_body"]
unlabelled_data.loc[unlabelled_data["body"].isnull(), "body"] = unlabelled_data.loc[unlabelled_data["body"].isnull(), "html_body_as_text"]

# Export to Pickle

# if only want bank-change-detail emails; 
# bankchange_emails = labelled_data[labelled_data.category_slug.str.contains("bank-detail-change") == True]
# bankchange_emails.to_pickle("data/Step0_Data/bank_change_emails_and_threads.pkl")
# print(len(bankchange_emails))

# labelled_data.to_pickle("data/Step0_Data/labelled_emails_and_threads.pkl")

unlabelled_data.to_pickle("data/Step0_Data/unlabelled_emails_and_threads_run6.pkl")

# del labelled_data, unlabelled_data

del unlabelled_data
