==> Step2_Assess_New_Features.ipynb <==
# Goal of script

Created on 01/03/2023 by Claire S

This script assesses the cleaned data in terms of split emails and new features created in the following scripts:

- Step0_Data_Extraction
- Step1_Regex_Data_Cleaning

This Script filters the data and tries to find more targets

%load_ext autoreload
%autoreload 2

import pandas as pd
import pickle
import numpy as np

# pandas options
pd.set_option('display.max_rows', 1000)
pd.set_option('display.max_columns', 1000)

# try and import my own functions:
import sys
import ipynb
import re
sys.path.append('/project/Xelix_Project/utils')

# go through this for each run
this_file = "data/Step1_Data/labelled_emails_Step1_Split_newfeatures_V2.pkl"
# this_file = "data/labelled_emails_more_features_checked.csv"

# go through this for each run
# df = pd.read_csv(this_file)
df = pd.read_pickle(this_file)

# df.to_csv('data/labelled_emails_V2.csv')

# df = pd.read_pickle("data/Step1_Data/unlabelled_emails_Step1_Split_newfeatures_run2.pkl")

print(len(df))

print(len(df.columns))
df.columns

## Some manual recoding of true bank change emails


bc_ids = [
186977,
186977,
282436,
283345,
284818,
287342,
291518,
291520,
292234,
292234,
292234,
349649,
351685,
356414,
380734,
380829
]

thread_ids = [
2,
1,
0,
0,
0,
0,
0,
0,
4,
3,
1,
0,
0,
0,
7,
2
]

matches = pd.DataFrame({'email_id':bc_ids,'num_in_thread':thread_ids})
matches['final_label'] = 1
matches

# Merge the two dataframes on the common columns
# df['final_label'] = 0
merged_df = df.merge(matches, on=['email_id', 'num_in_thread'], how='left')
merged_df['final_label'] = merged_df['final_label'].fillna(0).astype(int)

merged_df.columns

sum(merged_df['final_label'])

df = merged_df
sum(df['final_label'])

df.head(10)

df.info()

# Filter dataframe on conditions

df.head(1)

all_bchange = df.query('(final_label == 1)')
print(len(df.query('(final_label == 1)')))


# query = '''
# ((mentions_bank == True) | (mentions_bank_account == True) or (subject_mentions_bank == True)) & ((mentions_change == True) or (subject_mentions_change == True))
# '''
# new_df = df.query(query)
# print(len(new_df))

# query = '((mentions_bank) or (mentions_bank_account) or (subject_mentions_bank)) & ((mentions_change) or (subject_mentions_change))'
# new_df = df.query(query)
# print(len(new_df))
query = '((mentions_bank == True) or (mentions_bank_account == True) or (subject_mentions_bank == True)) & ((mentions_change == True) or (subject_mentions_change == True))'
new_df = df.query(query)
print(len(new_df))


# new_df = df.query('(mentions_bank == True) & (mentions_change == True)')
# print(len(new_df))

some_bchange = new_df.query('(final_label == 1)')
print(len(new_df.query('(final_label == 1)')))

i = 1
for row in all_bchange.itertuples():
    print("#" * 70, "\n\n")
    print(f'email index: {i}')
    i += 1
    print(f'subject line:\t{row.subject}')
    print(f'received from:\t{row.from_address}')
    print(f'is part of thread:\t{bool(row.is_part_of_thread)}')
    #print(f'num in thread:\t\t{int(row.num_in_thread)}')
    print(f'has attachment:\t\t{row.has_attachment}')
    print(f'has sort code:\t\t{row.has_sort_code}')
    print(f'mentions bank:\t\t{row.mentions_bank}')
    print(f'mentions b details:\t{row.mentions_account_specifics}')
    print(f'mentions change:\t{row.mentions_change}')
    print(f'subject mentions bank:\t{row.subject_mentions_bank}')
    print(f'subject mentions change:{row.subject_mentions_change}')
    #print(f'sent to:\t{row.sent_to}')
    print(f'sent to {row.sent_to_how_many} people')
    print(f'date sent:\t{row.internal_date}')
    print(f'slug:\t\t{row.category_slug}\n')
    print(row.email_text_body)

i = 1
for row in some_bchange.itertuples():
    print("#" * 70, "\n\n")
    print(f'email index: {i}')
    i += 1
    print(f'subject line:\t{row.subject}')
    print(f'received from:\t{row.from_address}')
    print(f'is part of thread:\t{bool(row.is_part_of_thread)}')
    #print(f'num in thread:\t\t{int(row.num_in_thread)}')
    print(f'has attachment:\t\t{row.has_attachment}')
    print(f'has sort code:\t\t{row.has_sort_code}')
    print(f'mentions bank:\t\t{row.mentions_bank}')
    print(f'mentions b details:\t{row.mentions_account_specifics}')
    print(f'mentions change:\t{row.mentions_change}')
    print(f'subject mentions bank:\t{row.subject_mentions_bank}')
    print(f'subject mentions change:{row.subject_mentions_change}')
    #print(f'sent to:\t{row.sent_to}')
    print(f'sent to {row.sent_to_how_many} people')
    print(f'date sent:\t{row.internal_date}')
    print(f'slug:\t\t{row.category_slug}\n')
    print(row.email_text_body)

## This was for manual labelling

new_df = new_df.assign(true_positives=pd.Series(1, index=new_df.index))
new_df = new_df.assign(false_positives=pd.Series(0, index=new_df.index))
new_df = new_df.assign(uncertain=pd.Series(0, index=new_df.index))
new_df = new_df.assign(notes=pd.Series(None, index=new_df.index,dtype=str))

# new_df['true_positives'] = 1
# new_df['false_positives'] = 0
# new_df['uncertain'] = 0
# new_df['notes'] = None

new_df.head(10)

# Export data

this_file_csv = 'data/Step2_Data/labelled_emails_filtered.csv'

new_df.to_csv(this_file_csv)

## Read more examples

i = 1
for row in new_df[0:100].itertuples():
    print("#" * 70, "\n\n")
    print(f'email index: {i}')
    i += 1
    print(f'subject line:\t{row.subject}')
    print(f'received from:\t{row.from_address}')
    print(f'is part of thread:\t{bool(row.is_part_of_thread)}')
    #print(f'num in thread:\t\t{int(row.num_in_thread)}')
    print(f'has attachment:\t\t{row.has_attachment}')
    print(f'has sort code:\t\t{row.has_sort_code}')
    print(f'mentions bank:\t\t{row.mentions_bank}')
    print(f'mentions b details:\t{row.mentions_account_specifics}')
    print(f'mentions change:\t{row.mentions_change}')
    #print(f'sent to:\t{row.sent_to}')
    print(f'sent to {row.sent_to_how_many} people')
    print(f'date sent:\t{row.internal_date}')
    print(f'slug:\t\t{row.category_slug}\n')
    print(f'{row.email_text_body}\n')

# new_df.to_csv('data/Step2_Data/initial_bank_change_emails.csv')

total_emails = 55199+56979+57668+131759+83226+38894
total_emails

initial_filter = 722+651+626+356+567+140
initial_filter

perc = (initial_filter*100)/total_emails
perc

total_emails/100 # 1 percent of all emails is:

# merge dataframes here

import pandas as pd

df1 = pd.read_csv("data/Step2_Data/pot_new_targets_run1.csv")
df2 = pd.read_csv("data/Step2_Data/pot_new_targets_run2.csv")
df3 = pd.read_csv("data/Step2_Data/pot_new_targets_run3.csv")
df4 = pd.read_csv("data/Step2_Data/pot_new_targets_run4.csv")
df5 = pd.read_csv("data/Step2_Data/pot_new_targets_run5.csv")
df6 = pd.read_csv("data/Step2_Data/pot_new_targets_run6.csv")

tot_df = pd.concat([df1,df2,df3,df4,df5,df6],ignore_index=True)



tot_df.to_csv('data/Step2_Data/merged_potential_targets_v2.csv')

## Save as pickle?

df = pd.read_pickle("bank_change_emails_split_newfeatures.pkl")

i = 1
for row in df[0:100].itertuples():
    print("#" * 70, "\n\n")
    print(f'email index: {i}')
    i += 1
    print(f'subject line:\t{row.subject}')
    print(f'received from:\t{row.from_address}')
    print(f'is part of thread:\t{bool(row.is_part_of_thread)}')
    #print(f'num in thread:\t\t{int(row.num_in_thread)}')
    print(f'has attachment:\t\t{row.has_attachment}')
    print(f'has sort code:\t\t{row.has_sort_code}')
    print(f'mentions bank:\t\t{row.mentions_bank}')
    print(f'mentions b details:\t{row.mentions_account_specifics}')
    print(f'mentions change:\t{row.mentions_change}')
    #print(f'sent to:\t{row.sent_to}')
    print(f'sent to {row.sent_to_how_many} people')
    print(f'date sent:\t{row.internal_date}')
    print(f'slug:\t\t{row.category_slug}\n')
    print(f'{row.email_text_body}\n')
